Separate classes into headers
